Improvements
-It all works
-Everything (or almost) is configurable in one config file (json presets went to trash)
-Configs include changing text colour and if the text drops a shadow

By default, the GTNH pack setup is in config but technically, with the new config, someone can use it as the normal CLS version. The exception is that the there can only be a background, a logo and two loading bars (main and material) set, no additional text or images for instance.
-CLS now turns the forge splash screen directly in configs